Update token contract

+++ contracts/HaikoinMarket.sol
@@ -20,7 +20,7 @@ contract HaikoinMarket is ReentrancyGuard {
   }
 
   struct MarketItem {
-    uint itemId;
+    uint256 itemId;
     address nftContract;
     uint256 tokenId;
     address payable seller;
@@ -31,8 +31,8 @@ contract HaikoinMarket is ReentrancyGuard {
 
   mapping(uint256 => MarketItem) private idToMarketItem;
 
-  event MarketItemCreated (
-    uint indexed itemId,
+  event MarketItemCreated(
+    uint256 indexed itemId,
     address indexed nftContract,
     uint256 indexed tokenId,
     address seller,
@@ -58,7 +58,7 @@ contract HaikoinMarket is ReentrancyGuard {
     _itemIds.increment();
     uint256 itemId = _itemIds.current();
 
-    idToMarketItem[itemId] =  MarketItem(
+    idToMarketItem[itemId] = MarketItem(
       itemId,
       nftContract,
       tokenId,
@@ -83,13 +83,17 @@ contract HaikoinMarket is ReentrancyGuard {
 
   /* Creates the sale of a marketplace item */
   /* Transfers ownership of the item, as well as funds between parties */
-  function createMarketSale(
-    address nftContract,
-    uint256 itemId
-    ) public payable nonReentrant {
-    uint price = idToMarketItem[itemId].price;
-    uint tokenId = idToMarketItem[itemId].tokenId;
-    require(msg.value == price, "Please submit the asking price in order to complete the purchase");
+  function createMarketSale(address nftContract, uint256 itemId)
+    public
+    payable
+    nonReentrant
+  {
+    uint256 price = idToMarketItem[itemId].price;
+    uint256 tokenId = idToMarketItem[itemId].tokenId;
+    require(
+      msg.value == price,
+      "Please submit the asking price in order to complete the purchase"
+    );
 
     idToMarketItem[itemId].seller.transfer(msg.value);
     IERC721(nftContract).transferFrom(address(this), msg.sender, tokenId);
@@ -101,14 +105,14 @@ contract HaikoinMarket is ReentrancyGuard {
 
   /* Returns all unsold market items */
   function fetchMarketItems() public view returns (MarketItem[] memory) {
-    uint itemCount = _itemIds.current();
-    uint unsoldItemCount = _itemIds.current() - _itemsSold.current();
-    uint currentIndex = 0;
+    uint256 itemCount = _itemIds.current();
+    uint256 unsoldItemCount = _itemIds.current() - _itemsSold.current();
+    uint256 currentIndex = 0;
 
     MarketItem[] memory items = new MarketItem[](unsoldItemCount);
-    for (uint i = 0; i < itemCount; i++) {
+    for (uint256 i = 0; i < itemCount; i++) {
       if (idToMarketItem[i + 1].owner == address(0)) {
-        uint currentId = i + 1;
+        uint256 currentId = i + 1;
         MarketItem storage currentItem = idToMarketItem[currentId];
         items[currentIndex] = currentItem;
         currentIndex += 1;
@@ -119,20 +123,20 @@ contract HaikoinMarket is ReentrancyGuard {
 
   /* Returns only items that a user has purchased */
   function fetchMyNFTs() public view returns (MarketItem[] memory) {
-    uint totalItemCount = _itemIds.current();
-    uint itemCount = 0;
-    uint currentIndex = 0;
+    uint256 totalItemCount = _itemIds.current();
+    uint256 itemCount = 0;
+    uint256 currentIndex = 0;
 
-    for (uint i = 0; i < totalItemCount; i++) {
+    for (uint256 i = 0; i < totalItemCount; i++) {
       if (idToMarketItem[i + 1].owner == msg.sender) {
         itemCount += 1;
       }
     }
 
     MarketItem[] memory items = new MarketItem[](itemCount);
-    for (uint i = 0; i < totalItemCount; i++) {
+    for (uint256 i = 0; i < totalItemCount; i++) {
       if (idToMarketItem[i + 1].owner == msg.sender) {
-        uint currentId = i + 1;
+        uint256 currentId = i + 1;
         MarketItem storage currentItem = idToMarketItem[currentId];
         items[currentIndex] = currentItem;
         currentIndex += 1;
@@ -143,20 +147,20 @@ contract HaikoinMarket is ReentrancyGuard {
 
   /* Returns only items a user has created */
   function fetchItemsCreated() public view returns (MarketItem[] memory) {
-    uint totalItemCount = _itemIds.current();
-    uint itemCount = 0;
-    uint currentIndex = 0;
+    uint256 totalItemCount = _itemIds.current();
+    uint256 itemCount = 0;
+    uint256 currentIndex = 0;
 
-    for (uint i = 0; i < totalItemCount; i++) {
+    for (uint256 i = 0; i < totalItemCount; i++) {
       if (idToMarketItem[i + 1].seller == msg.sender) {
         itemCount += 1;
       }
     }
 
     MarketItem[] memory items = new MarketItem[](itemCount);
-    for (uint i = 0; i < totalItemCount; i++) {
+    for (uint256 i = 0; i < totalItemCount; i++) {
       if (idToMarketItem[i + 1].seller == msg.sender) {
-        uint currentId = i + 1;
+        uint256 currentId = i + 1;
         MarketItem storage currentItem = idToMarketItem[currentId];
         items[currentIndex] = currentItem;
         currentIndex += 1;
